Replaced CA_STATE bit manipulation with a simple bool in MediaFoundation_Camera
Seriously, who in their right mind would use such a thing when a bool is enough?
Even if bool is not enough, there are always enums.
Another thing is if you actually need bitmasks where several flags can be set at the same time, but you still can use enums for that too.