Fixed and refactored capability getting/setting code in Media Foundation
It should be able to detect all of PixelFormats supported by a camera now.

In fact, it detects too much now. It detects 1 FPS capabilities, the ones that Direct Show marks as still image capabilities, which cause our Media Foundation code to crash when trying to capture using them. There was a similar issue with Direct Show, which was fixed by filtering still image capabilities out, using only video capturing ones. But here, in Media Foundation, it doesn't seem possible to differentiate between still image capture and video capture capabilities, they seem to be both marked as video capture ones, not still image capture (in contrast to Direct Show), and there doesn't seem to be any flag that would help us differentiate them. We could just filter out all capabilities with FPS=1, but this is more of a hack than a solution, since it's quite possible for some camera to support video capturing in such FPS, but we will be filtering it out, not allowing a library user to use it.